Java TCP GUI Chat
As a project in my studies I was asked to make a chat application.
The user who runs the Server JAR is considered as the server and all other users which are in the same WIFI as the server can connect to the chat and communicate along with advaced features that I made.
All the code here is made solely by me from scratch.